ImageBackdrop refactor
