+++ src/components/Config/index.tsx
@@ -6,16 +6,19 @@ import useSpotify from "@/modules/spotify/state";
 import { setSpotifyToken } from "@/modules/spotify/token";
 import { connectSpotifyPlayer } from "@/modules/spotify";
 import { progressGameFlow } from "@/modules/state";
+import ImageBackdrop from "../ImageBackdrop";
 
 const ConfigWrapper: React.FC<PropsWithChildren> = ({ children }) => {
 	return (
-		<div className="w-full">
+		<ImageBackdrop>
+			<div className="w-full">
 			<div className="mx-12 my-10">
-				<h1 className="font-bold">Remix Icebreaker Game</h1>
+				<h1 className="font-bold text-3xl">Remix</h1>
 				<hr className="my-2 border-gray-600 w-full" />
 				{children}
 			</div>
-		</div>
+			</div>
+		</ImageBackdrop>
 	);
 };
